refactor(updated subscription): refactored subscription updater function

+++ app/client/src/features/events/Questions/QuestionActions/QueueButton.tsx
@@ -55,41 +55,43 @@ export function QueueButton({ fragmentRef, currentQuestion }: QueueButtonProps) 
     }, [position]);
 
     const handleClick = () => {
-        commit({
-            variables: {
-                input: {
-                    questionId,
-                    eventId,
-                    adding: !isQueued
-                }
-            },
-            updater: (store: RecordSourceSelectorProxy) => {
-                const EventProxy = store.get(eventId);
-                if (!EventProxy) return;
-                const conn = ConnectionHandler.getConnection(EventProxy, 'QuestionQueueFragment_queuedQuestions');
-                const payload = store.getRootField('updateQuestionQueue');
-                if (!conn || !payload) return;
-                if (isQueued) {
-                    if (currentQuestion === position) {
-                        displaySnack('Cannot dequeue the current question');
-                    } else {
+        if (isQueued && currentQuestion === position) {
+            displaySnack('Cannot dequeue the current question');
+        } else if (isQueued && currentQuestion > position) {
+            displaySnack('Cannot dequeue already asked questions');
+        } else {
+            commit({
+                variables: {
+                    input: {
+                        questionId,
+                        eventId,
+                        adding: !isQueued
+                    }
+                },
+                updater: (store: RecordSourceSelectorProxy) => {
+                    const EventProxy = store.get(eventId);
+                    if (!EventProxy) return;
+                    const conn = ConnectionHandler.getConnection(EventProxy, 'QuestionQueueFragment_queuedQuestions');
+                    const payload = store.getRootField('updateQuestionQueue');
+                    if (!conn || !payload) return;
+                    if (isQueued) {
                         // If the question is being dequeued then delete the node from the connection
                         ConnectionHandler.deleteNode(conn, questionId);
+                    } else {
+                        const serverEdge = payload.getLinkedRecord('body');
+                        if (!serverEdge) return;
+                        const newEdge = ConnectionHandler.buildConnectionEdge(store, conn, serverEdge);
+                        if (!newEdge) return;
+                        ConnectionHandler.insertEdgeAfter(conn, newEdge);
                     }
-                } else {
-                    const serverEdge = payload.getLinkedRecord('body');
-                    if (!serverEdge) return;
-                    const newEdge = ConnectionHandler.buildConnectionEdge(store, conn, serverEdge);
-                    if (!newEdge) return;
-                    ConnectionHandler.insertEdgeAfter(conn, newEdge);
+                },
+                onCompleted: ({ updateQuestionQueue }) => {
+                    if (updateQuestionQueue.isError) {
+                        displaySnack(updateQuestionQueue.message);
+                    }
                 }
-            },
-            onCompleted: ({ updateQuestionQueue }) => {
-                if (updateQuestionQueue.isError) {
-                    displaySnack(updateQuestionQueue.message);
-                }
-            }
-        });
+            });
+        }
     };
     return (
         <Button
